chore: rename status-badge
